% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.meta.r
\name{apply.meta}
\alias{apply.meta}
\title{Applies a meta.data template onto an indicated variable.}
\usage{
apply.meta(variable = NA, template = NA)
}
\arguments{
\item{variable}{is the variable that we want to apply the template to}

\item{template}{is the template that we want to apply to the variable.}
}
\value{
variable with the meta.data attached.
}
\description{
Applies a meta.data template onto an indicated variable.
}
\examples{


iris$Sepal.Width <- create.meta(iris$Sepal.Width,
                                full.text = 'Variable indicating the Width of the Sepal',
                                value.labels = 'cm',
                                missing = 'NA' ,
                                comments = 'this is an example of more detailed meta.data 2')

write.meta <- function(variable) {
part.1 <-  as.data.frame(t(cbind(attributes(variable))))
return(part.1)}


meta.sepal.width <- write.meta(iris$Sepal.Width)



sepal.1 <- apply.meta(iris$Sepal.Width, meta.sepal.width )






}
