% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.metas.r
\name{apply.metas}
\alias{apply.metas}
\title{Applies meta.data.templates to a range of variables.}
\usage{
apply.metas(variable.set = NA, template.set = NA)
}
\arguments{
\item{variable.set}{is a data.frame containing multiple variables, which we want to apply meta.data to.}

\item{template.set}{is a set of templates, that we want to apply onto the variables indicated in variable.set.}
}
\value{
a data frame, containing the variables with the templates applied.
}
\description{
Applies meta.data.templates to a range of variables.
}
\examples{



iris$Sepal.Length <- create.meta(iris$Sepal.Length,
                                full.text = 'Variable indicating the length of the Sepal',
                                value.labels = 'cm',
                                missing = 'NA' ,
                                comments = 'this is an example of more detailed meta.data')


iris$Sepal.Width <- create.meta(iris$Sepal.Width,
                                full.text = 'Variable indicating the Width of the Sepal',
                                value.labels = 'cm',
                                missing = 'NA' ,
                                comments = 'this is an example of more detailed meta.data 2')

write.meta <- function(variable) {
part.1 <-  as.data.frame(t(cbind(attributes(variable))))
return(part.1)}


metas.1 <- rbind(write.meta(iris$Sepal.Length), write.meta(iris$Sepal.Width))



apply.metas(iris[,1:2], metas.1)



}
