% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.meta.R
\name{change.meta}
\alias{change.meta}
\title{Edits the meta.data of an indicated variable.}
\usage{
change.meta(
  variable = NA,
  long.label = NA,
  full.text = NA,
  value.labels = NA,
  missing = NA,
  associated.ids = NA,
  comments = NA,
  replace = F
)
}
\arguments{
\item{variable}{is a set of variables that you want to get the meta.data of}

\item{long.label}{is for indicating the long name of a variable}

\item{full.text}{is the full question of the text}

\item{value.labels}{indicates the values e.g., 'cm'}

\item{missing}{is missing indicator}

\item{associated.ids}{is a list of other ids associated with the variable.}

\item{comments}{is for anything else you want to include}

\item{replace}{T or F whether the original meta.data should be replaced.}
}
\value{
variable with the meta.data attached.
}
\description{
Edits the meta.data of an indicated variable.
}
\examples{

create.meta <- function( variable =NA , long.label =NA , full.text = NA ,
 value.labels =NA , missing =NA , associated.ids =NA , comments =NA ){
 original <- variable #make a copy
 part.1 <-  as.list(strsplit(as.character(substitute(variable)), '$'))
 attributes(original)$variable.label <- part.1[[length(part.1)]]
 #the rest are just normal comments etc.
 attributes(original)$long.label	 <- c(	long.label	)
 attributes(original)$full.text	 <- c(	full.text	)
 attributes(original)$value.labels	 <- c(	value.labels	)
 attributes(original)$missing	 <- c(	missing	)
 attributes(original)$associated.ids	 <- c(associated.ids)
 attributes(original)$comments <- c(	comments)
  return(original)}

iris$Sepal.Length <- create.meta(iris$Sepal.Length,
full.text = 'Variable indicating the length of the Sepal',
long.label = 'original',
value.labels = 'cm',
missing = 'NA' ,
comments = 'this is an example of more detailed meta.data')

change.meta(iris$Sepal.Length, long.label = 'test.1',
        full.text = 'this is another test of something',
        associated.ids = 'this is new one',
        replace = TRUE)
}
