% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.meta.r
\name{create.meta}
\alias{create.meta}
\title{Creates a new meta.data template for one particular variable.}
\usage{
create.meta(
  variable = NA,
  variable.label = NA,
  long.label = NA,
  full.text = NA,
  value.labels = NA,
  missing = NA,
  associated.ids = NA,
  comments = NA
)
}
\arguments{
\item{variable}{a target variable.}

\item{long.label}{a long label for the variable.}

\item{full.text}{full text of the question}

\item{value.labels}{the value labels in the variable (e.g., cm)}

\item{missing}{missing indicator}

\item{associated.ids}{these are ids for e.g., a dataset, a scale, or etc.}

\item{comments}{anything else that does not have a space so far.}
}
\value{
a variable with the attributes attached.
}
\description{
Creates a new meta.data template for one particular variable.
}
\examples{
iris$Sepal.Length <- create.meta(iris$Sepal.Length,
 full.text = 'Variable indicating the length of the Sepal',
 long.label = 'this indicates a variable measuring the length of the sepal',
 value.labels = 'cm' ,
 missing = 'NA' ,
 associated.ids = 'doi, link to iris data.set, links to other tools etc',
 comments = 'this is an example of more detailed meta.data')



iris$Sepal.Width <- create.meta(iris$Sepal.Width,
 full.text = 'Variable indicating the Width of the Sepal',
 value.labels = 'cm',
 missing = 'NA' ,
 comments = 'this is an example of more detailed meta.data 2')





}
