% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.metas.r
\name{write.metas}
\alias{write.metas}
\title{Writes a meta.data file for any range of variables e.g., a scale or dataset,}
\usage{
write.metas(variable.set)
}
\arguments{
\item{variable.set}{is a data.frame containing multiple variables, which we want to get the meta.data of.}
}
\value{
a data frame, containing all of the attributes from each of the variables in the data.set.
}
\description{
Writes a meta.data file for any range of variables e.g., a scale or dataset,
}
\examples{

iris$Sepal.Length <- create.meta(iris$Sepal.Length,
   full.text = 'Variable indicating the length of the Sepal',
   value.labels = 'cm',
   missing = 'NA' ,
   comment = 'this is an example of more detailed meta.data')

iris$Sepal.Width <- create.meta(iris$Sepal.Width,
   full.text = 'Variable indicating the Width of the Sepal',
   value.labels = 'cm',
   missing = 'NA' ,
   comment = 'this is an example of more detailed meta.data 2')

iris$Petal.Length <- create.meta(iris$Petal.Length,
   full.text = 'Variable indicating the length of the petal',
   value.labels = 'cm',
   missing = 'NA' ,
   comment = 'this is an example of more detailed meta.data 3 ')
iris$Petal.Width <- create.meta(iris$Petal.Width,
   full.text = 'Variable indicating the width of the petal',
   value.labels = 'cm',
   missing = 'NA' ,
   comment = 'this is an example of more detailed meta.data 4 ')

iris$Species  <- create.meta(iris$Species,
  full.text = 'Variable indicating the length of the Sepal',
  value.labels = 'cm',
  missing = 'NA' ,
  comment = 'this is an example of more detailed meta.data 5 ')


write.metas(iris)

}
